requirements.txt
python-dotenv==1.0.0
requests==2.31.0
websocket-client==1.6.4
eth-account==0.10.0
web3==6.15.1
pydantic==2.5.3
python-dateutil==2.8.2
python-telegram-bot==20.7
